Build && Autoresolve

+++ src/units/Infantry.java
@@ -10,38 +10,38 @@ public class Infantry extends Unit {
 
 	public void attack(Unit target) throws FriendlyFireException {
 		super.attack(target);
-		double c = target.getCurrentSoldierCount();
+		int c = target.getCurrentSoldierCount();
 		if (target instanceof Archer) {
 			if (this.getLevel() == 1) {
-				c -= (0.3 * this.getCurrentSoldierCount());
+				c -= (int)(0.3 * this.getCurrentSoldierCount());
 			}
 			if (this.getLevel() == 2) {
-				c -= (0.4 * this.getCurrentSoldierCount());
+				c -=(int) (0.4 * this.getCurrentSoldierCount());
 			}
 			if (this.getLevel() == 3) {
-				c -= (0.5 * this.getCurrentSoldierCount());
+				c -= (int)(0.5 * this.getCurrentSoldierCount());
 			}
 		}
 		if (target instanceof Infantry) {
 			if (this.getLevel() == 1) {
-				c -= (0.1 * this.getCurrentSoldierCount());
+				c -= (int)(0.1 * this.getCurrentSoldierCount());
 			}
 			if (this.getLevel() == 2) {
-				c -= (0.2 * this.getCurrentSoldierCount());
+				c -= (int)(0.2 * this.getCurrentSoldierCount());
 			}
 			if (this.getLevel() == 3) {
-				c -= (0.3 * this.getCurrentSoldierCount());
+				c -= (int)(0.3 * this.getCurrentSoldierCount());
 			}
 		}
 		if(target instanceof Cavalry) {
 			if (this.getLevel() == 1) {
-				c -= (0.1 * this.getCurrentSoldierCount());
+				c -= (int)(0.1 * this.getCurrentSoldierCount());
 			}
 			if (this.getLevel() == 1) {
-				c -= (0.2 * this.getCurrentSoldierCount());
+				c -= (int)(0.2 * this.getCurrentSoldierCount());
 			}
 			if (this.getLevel() == 1) {
-				c -= (0.25 * this.getCurrentSoldierCount());
+				c -= (int)(0.25 * this.getCurrentSoldierCount());
 			}
 		}
 		target.setCurrentSoldierCount((int)c);
